Copy my custom RAM settings to my own target
NT5CC128M16IP-DII is currenltly not surrported by litedram as I guess I
am the only user of litex that has a board with that chip on it.

Unfortunately it looks like Enclustras's example use tmings that matches
the "800" speedgrade timings, so I use that here as well as it appears
to work more stable than the 1600 timings I have currently defined.

Putting this here makes it such that only one custom repo needs to be
checked out.