Fix SDRAM reading! Now ith 32 bit width
The RAM in the simulation is x16 RAM, so is the NT5CC128M16 we have on
the Marx AX3, so this should be good. The issue was that we updated the
address at every clock cycle, but we needed to hold it for two to get
the correct bytes output.

On top of that the addressing needs to compensate for the data width as
well.